requirements.txt
boto3
PyMuPDF
chromadb
sentence-transformers
python-dotenv
openai